+++ packages/web-components/src/components/content-item-row/content-item-row-media-video.ts
@@ -11,7 +11,7 @@ import { html } from 'lit';
 import { property } from 'lit/decorators.js';
 import settings from '@carbon/ibmdotcom-utilities/es/utilities/settings/settings.js';
 import { carbonElement as customElement } from '@carbon/web-components/es/globals/decorators/carbon-element.js';
-import C4DVideoPlayerContainer from '../video-player/video-player-container';
+import C4DVideoPlayerContainer from '../video-player-v7/video-player-container';
 import styles from './content-item-row-media.scss';
 
 const { stablePrefix: c4dPrefix } = settings;